+++ mockredis/redis.py
@@ -30,6 +30,8 @@ class MockRedis(object):
         self.strict = strict
         # The 'Redis' store
         self.redis = defaultdict(dict)
+        # The 'PubSub' store
+        self.pubsub = defaultdict(list)
         self.timeouts = defaultdict(dict)
         # Dictionary from script to sha ''Script''
         self.shas = dict()
@@ -197,6 +199,7 @@ class MockRedis(object):
 
     def flushdb(self):
         self.redis.clear()
+        self.pubsub.clear()
         self.timeouts.clear()
 
     #### String Functions ####
@@ -900,6 +903,11 @@ class MockRedis(object):
             return args[:3] + args[4:]
 
         return args
+
+    #### PubSub commands ####
+
+    def publish(self, channel, message):
+        self.pubsub[channel].append(message)
 
     #### Internal ####